Use nom to parse WebP's container format
This causes quite an infectious macro_use at the top-level of the library, but that's unavoidable for now. It might make sense to move all of webp into its own crate in future, but if other parsers will ever use nom, it's not worthwhile. Nom itself is marked as an optional dependency, for the "webp" feature.

Add some test images from the WebP website, and register the WebP decoder for tests.

Right now, the old VP8 decoder is still used for VP8 image data, and it decodes it to luma only. In lossy-with-lossless-alpha files the alpha is ignored. Lossless files are unsupported and fail in the correct way. Animated files are unsupported and won't even parse yet. Color profiles and metadata are ignored.